requirements.txt
astroid==3.3.8
dill==0.3.9
isort==6.0.0
mccabe==0.7.0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
py4j==0.10.9.7
pylint==3.3.4
pyspark==3.5.4
pytesseract==0.3.13
tomlkit==0.13.2
